rename the preview modules to simulator module
