+++ packages/slice-machine/src/redux/saga.ts
@@ -1,8 +1,8 @@
 import { fork } from "redux-saga/effects";
 
-import { watchPreviewSagas } from "@src/modules/preview";
+import { watchSimulatorSagas } from "@src/modules/simulator";
 
 // Single entry point to start all Sagas at once
 export default function* rootSaga() {
-  yield fork(watchPreviewSagas);
+  yield fork(watchSimulatorSagas);
 }
