+++ packages/slice-machine/src/redux/type.ts
@@ -2,7 +2,7 @@ import { ModalStoreType } from "@src/modules/modal/types";
 import { LoadingStoreType } from "@src/modules/loading/types";
 import { UserContextStoreType } from "@src/modules/userContext/types";
 import { EnvironmentStoreType } from "@src/modules/environment/types";
-import { SimulatorStoreType } from "@src/modules/preview/types";
+import { SimulatorStoreType } from "@src/modules/simulator/types";
 
 export type SliceMachineStoreType = {
   modal: ModalStoreType;
